requirements.txt
flask
pandas
flask-cors
numpy
scikit-learn
geopy
lightgbm
requests
python-dotenv